koda-open-api-source/requirements.txt
# Mulberry Open API - Python Dependencies

# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0

# Database
pymongo==4.6.1
redis==5.0.1

# Auth & Security
cryptography==42.0.0
pyjwt==2.8.0
passlib==1.7.4
python-jose[cryptography]==3.3.0

# Validation
marshmallow==3.20.1
pydantic==2.5.3

# HTTP Client
requests==2.31.0

# Monitoring
prometheus-flask-exporter==0.23.0

# Production Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2
pytest-mock==3.12.0

# Utils
python-dotenv==1.0.0
python-dateutil==2.8.2

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
